talk_to_docs_full_project/fastapi/requirements.txt
fastapi
uvicorn
PyMuPDF
pytesseract
transformers
torch
Pillow